Restric access controler

+++ app/controllers/articles_controller.rb
@@ -31,25 +31,38 @@ class ArticlesController < ApplicationController
   end
 
 
-  def edit; end
+  def edit
+    unless @article.user == current_user
+      flash[:alert] = "You can only edit your own article."
+      redirect_to article_path
+    end
+  end
 
   def update
 
-    if @article.update(article_params)
-      flash[:succes] = 'Article has been updated'
-      redirect_to @article
+    unless @article.user == current_user
+      flash[:danger] = "You can only edit your own article."
+      redirect_to article_path
     else
-      flash.now[:danger] = 'Article has not been updated'
-      render :edit
+      if @article.update(article_params)
+        flash[:succes] = 'Article has been updated'
+        redirect_to @article
+      else
+        flash.now[:danger] = 'Article has not been updated'
+        render :edit
+      end
     end
   end
 
   def destroy
-
-    if @article.destroy
-      flash[:succes] = 'Article has been deleted'
-      redirect_to articles_path
-
+    unless @article.user == current_user
+      flash[:danger] = "You can only delete your own article."
+      redirect_to article_path
+    else
+      if @article.destroy
+        flash[:success] = 'Article has been deleted'
+        redirect_to articles_path
+      end
     end
   end
